refactor(utils): move the prometheus parser out of core, it depends on nothing
The module is stdlib-only text parsing — no sparkrun imports at all —
which is the shape of a utils leaf rather than core domain logic. Its two
consumers (core.monitoring, orchestration.nv_monitor) already imported it
lazily; those imports move with it.
